repo init: a minimalistic stack-based task cli
I've initialized a new repostory for this future project.

The main idea behind this cli is that you would have two memory 'blocks'
which are just 512 byte binary files that a split into two stacks,
with the first section a fourth the size of the block.

The first section/stack is a stack that stores the sizes in u16 of each
of the tasks stored in the second stack. And the second stack, like said
before, contains the actual contents of the tasks.

When the block is pushed to, the size of task is first stored in the
first stack and then the binary contents of it is stored in the second.
Incrementing the respective stack pointers (stored at the start of the
file) upon pushes and decrementing on pops. The location of the second
stack is stored immediately after the stack pointers.

There are two memory blocks in the program's home dir, one for short
term tasks and one for long term tasks.

Stack overflows are a defined and intended behaviour by the taskstack
CLI. As they are checked for before any pushes, and an appropriete
randomly chosen message from the done manifesto is written. As if you're
overflowing the stack for tasks, you should probably get something done.

This is done so that you are compelled to get things done, or to just
pop them off the stack. Because even abandoning something is still
getting it done, or well, being done with it and moving on. It also
fixes the order in which things can be done so you are forced to do it
or abandon it if you want to move on from it.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>