v0.1.0: written all the basic core functionality
I've written all the core functionality as specified in the spec
(first commit).

The cli takes standard form and the project is structured in a way
so that library-ish functions are in the root while cli central
functions and code are stored in the main.zig file.

The program works with 'memory blocks' which are just 512 byte files
one for each short-term and long-term memory-blocks. Each memory block
consists of two stacks, the size stack and the contents stack. The size
stack stores the size of the contents as a u16, while the contents stack
, well, stores the contents of the stack entry.

I've implemented all possible cli commands and options.

The only next possible step that I could take is through adding dates
to each of the stack entries as a simple utc u64 or something similar to
that. I could also expand the memory blocks to 1KiB if necessary.

The u64 could be implemented through an 'addTimestamp' function that
takes in a pointer to an array pointer.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>