v0.2.0: added timestamps to tasks
I've added timestamps to the start of the contents of tasks.
The timestamps are stored as an big endian i64 (8 bytes) and they begin
every single task, so what the program does when pushing tasks is it
inserts the timestamps at the start of the content and when reading the
contents (when listing or popping) the time is stripped from the start
of the slice and processed separately.

Signed-off-by: Ethan Ma <kalscium@protonmail.com>